chg: trying to get rid of code duplication a bit
